+++ packages/porygon-typeorm/CHANGELOG.md
@@ -20,3 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [UNRELEASED]
+
+### Added
+- Added `CrudTypeOrmModuleType` enum.
+- Added `TypeOrmCrudContainerModuleApi` module.
+- Added `QueryToFindQueryTypeOrmConverter` type.